third_party/WebKit/LayoutTests/fast/forms/label/contenteditable-label-focus-expected.txt

Calling target.focus()
focus received by [object HTMLLabelElement]
PASS document.activeElement is target
PASS successfullyParsed is true

TEST COMPLETE
I'm editable. Just a button
